Move ManageUserButtons to its own file
